Renaming CLI flags, addressing other comments
