+++ pkg/controller/appdefinition/jobs_test.go
@@ -3,6 +3,7 @@ package appdefinition
 import (
 	"testing"
 
+	"github.com/acorn-io/acorn/pkg/controller/namespace"
 	"github.com/acorn-io/acorn/pkg/scheme"
 	"github.com/acorn-io/baaah/pkg/router/tester"
 )
@@ -16,7 +17,7 @@ func TestJobsLabels(t *testing.T) {
 }
 
 func TestJobsLabelsNamespace(t *testing.T) {
-	tester.DefaultTest(t, scheme.Scheme, "testdata/job/labels-namespace", AddNamespace)
+	tester.DefaultTest(t, scheme.Scheme, "testdata/job/labels-namespace", namespace.AddNamespace)
 }
 
 func TestCronJobs(t *testing.T) {
