+++ pkg/controller/appdefinition/namespace_test.go
@@ -3,6 +3,7 @@ package appdefinition
 import (
 	"testing"
 
+	"github.com/acorn-io/acorn/pkg/controller/namespace"
 	"github.com/acorn-io/acorn/pkg/scheme"
 	"github.com/acorn-io/baaah/pkg/router/tester"
 )
@@ -16,9 +17,9 @@ func TestAssignTargetNamespace(t *testing.T) {
 }
 
 func TestLabelsAnnotationsBasic(t *testing.T) {
-	tester.DefaultTest(t, scheme.Scheme, "testdata/propagation_basic", AddNamespace)
+	tester.DefaultTest(t, scheme.Scheme, "testdata/propagation_basic", namespace.AddNamespace)
 }
 
 func TestLabelsAnnotationsNoConfigset(t *testing.T) {
-	tester.DefaultTest(t, scheme.Scheme, "testdata/propagation_noconfig", AddNamespace)
+	tester.DefaultTest(t, scheme.Scheme, "testdata/propagation_noconfig", namespace.AddNamespace)
 }
