+++ pkg/controller/appdefinition/testdata/propagation_basic/existing.yaml
@@ -1,6 +1,6 @@
 apiVersion: v1
 data:
-  config: '{"annotationsPropagation":["foo"], "labelsPropagation":["bar"]}'
+  config: '{"propagateProjectAnnotations":["foo"], "propagateProjectLabels":["bar"]}'
 kind: ConfigMap
 metadata:
   name: acorn-config
